+++ src/main/java/com/shoppingcartservice/sachin/Entities/Product/Product.java
@@ -6,7 +6,7 @@ import java.util.List;
 @Entity
 public class Product {
     @Id@GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long productId;
+    private int productId;
     private String name;
     private double price;
     private String details;
@@ -17,11 +17,11 @@ public class Product {
 
 
 
-    public long getProductId() {
+    public int getProductId() {
         return productId;
     }
 
-    public void setProductId(long productId) {
+    public void setProductId(int productId) {
         this.productId = productId;
     }
